maildefender/requirements.txt
rich
python-whois
dnspython
requests
beautifulsoup4